Residual smoothing refactor --> making utilities directory broken apart.
