+++ src/curvefit/core/utils.py
@@ -1,6 +1,5 @@
 import numpy as np
 import pandas as pd
-from xspline import XSpline
 from copy import deepcopy
 from collections import OrderedDict
 from curvefit.core.functions import *
@@ -95,96 +94,6 @@ def get_derivative_of_column_in_ln_space(df, col_obs, col_t, col_grp):
     # combine all the data frames
     df_result = pd.concat([df_all[g] for g in groups])
     return df_result
-
-
-def local_smoother(df,
-                   col_val,
-                   col_axis,
-                   radius=None):
-    """Compute the neighbor mean and std of the residual matrix.
-
-    Args:
-        df (pd.DataFrame): Residual data frame.
-        col_val ('str'): Name for column that store the residual.
-        col_axis (list{str}): List of two axis column names.
-        radius (list{int} | None, optional):
-            List of the neighbor radius for each dimension.
-
-    Returns:
-        pd.DataFrame:
-            Return the data frame with two extra columns contains neighbor
-            mean and std.
-    """
-    radius = [0, 0] if radius is None else radius
-    assert col_val in df
-    assert len(col_axis) == 2
-    assert len(radius) == 2
-    assert all([col in df for col in col_axis])
-    assert all([r >= 0 for r in radius])
-
-    col_mean = '_'.join([col_val, 'mean'])
-    col_std = '_'.join([col_val, 'std'])
-
-    # group by the axis
-    df = df.groupby(col_axis, as_index=False).agg({
-        col_val: [np.sum, lambda x: np.sum(x**2), 'count']
-    })
-
-    col_sum = '_'.join([col_val, 'sum'])
-    col_sum2 = '_'.join([col_val, 'sum2'])
-    col_count = '_'.join([col_val, 'count'])
-
-    df.columns = df.columns.droplevel(1)
-    df.columns = list(df.columns[:-3]) + [col_sum, col_sum2, col_count]
-
-    sum_mat, indices, axis = df_to_mat(df, col_val=col_sum, col_axis=col_axis,
-                                       return_indices=True)
-    sum2_mat = df_to_mat(df, col_val=col_sum2, col_axis=col_axis)
-    count_mat = df_to_mat(df, col_val=col_count, col_axis=col_axis)
-
-    sum_vec = convolve_sum(sum_mat, radius)[indices[:, 0], indices[:, 1]]
-    sum2_vec = convolve_sum(sum2_mat, radius)[indices[:, 0], indices[:, 1]]
-    count_vec = convolve_sum(count_mat, radius)[indices[:, 0], indices[:, 1]]
-
-    df[col_mean] = sum_vec/count_vec
-    df[col_std] = np.sqrt(sum2_vec/count_vec - df[col_mean]**2)
-    df.drop(columns=[col_sum, col_sum2, col_count], inplace=True)
-
-    return df
-
-
-def neighbor_mean_std(df,
-                      col_val,
-                      col_group,
-                      col_axis,
-                      radius=None):
-    """Compute the neighbor mean and std of the residual matrix.
-
-    Args:
-        df (pd.DataFrame): Residual data frame.
-        col_val ('str'): Name for column that store the residual.
-        col_group ('str'): Name for column that store the group label.
-        col_axis (list{str}): List of two axis column names.
-        radius (list{int} | None, optional):
-            List of the neighbor radius for each dimension.
-
-    Returns:
-        pd.DataFrame:
-            Return the data frame with two extra columns contains neighbor
-            mean and std.
-    """
-    assert col_group in df
-    groups = df[col_group].unique()
-
-    df_list = []
-    for i, group in enumerate(groups):
-        df_sub = df[df[col_group] == group].reset_index(drop=True)
-        df_result = local_smoother(df_sub, col_val, col_axis,
-                                   radius=radius)
-        df_result[col_group] = group
-        df_list.append(df_result)
-
-    return pd.concat(df_list)
 
 
 # TODO: replace by the data translator?
@@ -564,73 +473,6 @@ def truncate_draws(t, draws, draw_space, last_day, last_obs, last_obs_space):
         final_draws = final_draws.ravel()
 
     return final_draws
-
-
-def convolve_sum(mat, radius=None):
-    """Convolve sum a 2D matrix by given radius.
-
-    Args:
-        mat (numpy.ndarray):
-            Matrix of interest.
-        radius (arraylike{int} | None, optional):
-            Given radius, if None assume radius = (0, 0).
-
-    Returns:
-        numpy.ndarray:
-            The convolved sum, with the same shape with original matrix.
-    """
-    mat = np.array(mat).astype(float)
-    assert mat.ndim == 2
-    if radius is None:
-        return mat
-    assert hasattr(radius, '__iter__')
-    radius = np.array(radius).astype(int)
-    assert radius.size == 2
-    assert all([r >= 0 for r in radius])
-    # import pdb; pdb.set_trace()
-    shape = np.array(mat.shape)
-    window_shape = tuple(radius*2 + 1)
-
-    mat = np.pad(mat, ((radius[0],),
-                       (radius[1],)), 'constant', constant_values=np.nan)
-    view_shape = tuple(np.subtract(mat.shape, window_shape) + 1) + window_shape
-    strides = mat.strides*2
-    sub_mat = np.lib.stride_tricks.as_strided(mat, view_shape, strides)
-    sub_mat = sub_mat.reshape(*shape, np.prod(window_shape))
-
-    return np.nansum(sub_mat, axis=2)
-
-
-def df_to_mat(df, col_val, col_axis, return_indices=False):
-    """Convert columns in data frame to matrix.
-
-    Args:
-        df (pandas.DataFrame): Given data frame.
-        col_val (str): Value column.
-        col_axis (list{str}): Axis column.
-        return_indices (bool, optional):
-            If True, return indices of the original values and the corresponding
-            axis values in the data frame.
-
-    Returns:
-        numpy.ndarray: Converted matrix.
-    """
-    assert col_val in df
-    assert all([c in df for c in col_axis])
-
-    vals = df[col_val].values
-    axis = df[col_axis].values.astype(int)
-    indices = (axis - axis.min(axis=0)).astype(int)
-    shape = tuple(indices.max(axis=0).astype(int) + 1)
-
-    mat = np.empty(shape)
-    mat.fill(np.nan)
-    mat[indices[:, 0], indices[:, 1]] = vals
-
-    if return_indices:
-        return mat, indices, axis
-    else:
-        return mat
 
 
 def smooth_draws(mat, radius=0, sort=False):
